--- src/views/dialogs/DraftDialog/style.css.ts
@@ -1,48 +1,0 @@
-import { colors } from '@sopt-makers/colors';
-import { style } from '@vanilla-extract/css';
-
-import { theme } from 'styles/theme.css';
-
-export const mainText = style({
-  color: colors.gray950,
-  ...theme.font.HEADING_5_20_B,
-});
-
-export const subText = style({
-  marginTop: 2,
-  color: colors.gray300,
-  ...theme.font.BODY_2_16_M,
-});
-
-export const buttonOutside = style({
-  width: 'fit-content',
-  borderRadius: 12,
-  color: colors.white,
-  backgroundColor: '#BDEC00',
-  marginTop: 20,
-  marginLeft: 'auto',
-});
-
-export const buttonInside = style({
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-  width: 'fit-content',
-  padding: '13px 20px',
-  borderRadius: 12,
-  transition: 'background-color 0.3s ease-out',
-  ...theme.font.LABEL_3_14_SB,
-
-  color: colors.white,
-  backgroundColor: '#BDEC00',
-
-  ':hover': {
-    backgroundColor: '#99BF00',
-  },
-  ':active': {
-    margin: '0 auto',
-    borderRadius: '100%',
-    width: '50%',
-    whiteSpace: 'nowrap',
-  },
-});
