+++ src/views/dialogs/ExitDialog/index.tsx
@@ -2,7 +2,7 @@ import { forwardRef } from 'react';
 
 import Dialog from '@components/Dialog';
 
-import { buttonInside, buttonOutside, buttonWrapper, mainText, subText } from './style.css';
+import { buttonInside, buttonOutside, buttonWrapper, mainText, subText } from '../style.css';
 
 const ExitDialog = forwardRef<HTMLDialogElement>((_, ref) => {
   return (
